backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-jose
geopy
qrcode[pil]
razorpay
